refactor: CLI did resolve
