+++ packages/daf-cli/src/did.ts
@@ -246,3 +246,16 @@ did
       console.error(e)
     }
   })
+
+did
+  .command('resolve <didUrl>')
+  .description('Resolve DID Document')
+  .action(async (didUrl) => {
+    const agent = getAgent(program.config)
+    try {
+      const ddo = await agent.resolveDid({ didUrl })
+      console.log(ddo)
+    } catch (e) {
+      console.error(e.message)
+    }
+  })
